* Changed the path structure of the VSS project. For single project to be imported default behavior is to put all its contents directly into the SVN folder selected by user. For multiple projects (or single if option is selected) project names are created as folders under SVN folder selected by user. The option to include VSS project subfolders is removed (always ON now). * “Update revision properties” and “Remove VSS binding” are no longer options and are always performed * Fixed a bug the revision merging logic and changed it a little. It now merges consecutive revisions if the user is the same, the comment is the same (even if empty) and the time span between them is 5 seconds (configurable in advanced settings) * Versions where branching had occurred are ignored * Pinned files are detected and user is warned about them in the end. TO DO: figure out the version they are pinned to and update to that version * Revision properties are updated remotely now – local access to SVN is no longer required. The original VSS username and time are set for each revision, not the SVN logged-in user. This requires pre-revprop-change hook to be set-up on server. * Made sure VSS bindings are removed for projects from VS 2003 to 2012. Also added VC 6 dsp and dsw support * Fixed a bug with VSS project selection dialog (project can actually be unselected now) * Added a button to create SVN folder for the root project * Got rid of reading command line arguments at the start. All settings are automatically saved on exit and loaded at the start of the program. The settings are saved into XML file in the same folder as the program. * Improved GUI, progress, and log output

+++ Cyotek.SourceSafeSvnMigration/ProgressEventArgs.cs
@@ -2,40 +2,40 @@
 
 namespace Cyotek.SourceSafeSvnMigration
 {
-  public class ProgressEventArgs : EventArgs
-  {
-  #region  Public Constructors  
+	public class ProgressEventArgs : EventArgs
+	{
+		#region  Public Constructors
 
-    public ProgressEventArgs(int percentComplete)
-      : this(null, percentComplete)
-    { }
+		public ProgressEventArgs(double percentComplete)
+			: this(null, percentComplete)
+		{ }
 
-    public ProgressEventArgs(string status)
-      : this(status, 0)
-    { }
+		public ProgressEventArgs(string status)
+			: this(status, 0.0)
+		{ }
 
-    public ProgressEventArgs(string status, int percentComplete)
-      : this()
-    {
-      this.Status = status;
-      this.PercentComplete = percentComplete;
-    }
+		public ProgressEventArgs(string status, double percentComplete)
+			: this()
+		{
+			this.Status = status;
+			this.PercentComplete = percentComplete;
+		}
 
-  #endregion  Public Constructors  
+		#endregion  Public Constructors
 
-  #region  Protected Constructors  
+		#region  Protected Constructors
 
-    protected ProgressEventArgs()
-    { }
+		protected ProgressEventArgs()
+		{ }
 
-  #endregion  Protected Constructors  
+		#endregion  Protected Constructors
 
-  #region  Public Properties  
+		#region  Public Properties
 
-    public int PercentComplete { get; protected set; }
+		public double PercentComplete { get; protected set; }
 
-    public string Status { get; protected set; }
+		public string Status { get; protected set; }
 
-  #endregion  Public Properties  
-  }
+		#endregion  Public Properties
+	}
 }
